modified script
- beautified the code for easier readability
